Make recommendations hash favour player
Arguments specified were previously unclear that the recommendations
are in favour for a player in a given player, which also created a
potential bug in the code in generation the recommended move.

Now it should be clearer that the recommendations hash is for a given
board, and a given player's symbol, a recommendation of which index
board index should be chosen.